pyproject.toml
[project]
name = "smartcities-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "deep-translator>=1.11.4",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.117.1",
    "fastmcp>=2.12.3",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-gemini>=0.1.1",
    "langchain-google-genai>=2.1.12",
    "langchain-mcp-adapters>=0.1.10",
    "langchain-ollama>=0.3.8",
    "langchain-text-splitters>=0.3.11",
    "langgraph>=0.6.7",
    "mcp>=1.14.1",
    "ollama>=0.5.4",
    "playwright>=1.55.0",
    "pydantic>=2.11.9",
    "pypdf>=6.1.0",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
]
